requirements.txt
slackclient>=2.6.2
python-dotenv>=0.13.0